Create info struct which will help in the future
This struct unifies the information about the stack, like the pointers
to them, their sizes, their max, mid and min values (currently).
The reason of creating this struct is to group the useful information
that will help me sort the stack, without changing the whole code in
case that I want to expand it (adding more information).